+++ packs/sensu/etc/st2_handler.py
@@ -27,8 +27,8 @@ except ImportError:
 
 # ST2 configuration
 
-ST2_API_BASE_URL = 'https://localhost/api/v1/'
-ST2_AUTH_BASE_URL = 'https://localhost/auth/v1/'
+ST2_API_BASE_URL = None  # 'https://localhost/api/v1/'
+ST2_AUTH_BASE_URL = None  # 'https://localhost/auth/v1/'
 ST2_USERNAME = None
 ST2_PASSWORD = None
 ST2_API_KEY = None
@@ -240,7 +240,7 @@ def _post_webhook(url, body, verbose=False):
         status = r.status_code
 
         if status in UNREACHABLE_CODES:
-            msg = 'Webhook URL %s does not exist. Check StackStorm installation!'
+            msg = 'Webhook URL %s does not exist. Check StackStorm installation!' % (url)
             raise Exception(msg)
 
         if status not in OK_CODES: