+++ packs/nagios/CHANGES.md
@@ -3,3 +3,8 @@
 ## v0.1.0
 
 * Initial release
+
+## v0.2.0
+
+* Update the event handler script with new API and auth URI
+* Add verbose mode to the event handler script (debbuging only)